revert: keep src/{esi,hourly,refresh,sso,supabase} as JS
Per request, revert the JS-to-TS conversion of the standalone scripts.
Restores tsconfig allowJs:true, the node-run npm scripts, and drops
tsx and @types/jsonwebtoken devDependencies.
