+++ package.json
@@ -23,10 +23,12 @@
   },
   "scripts": {
     "build": "next build",
-    "hourly": "tsx src/hourly.ts",
-    "refresh": "tsx src/refresh.ts",
+    "connect": "node src/connect.js",
     "dev": "next dev",
+    "hourly": "node src/hourly.js",
+    "refresh": "node src/refresh.js",
     "lint": "eslint .",
+    "ping": "node src/ping-db.js",
     "prepare": "husky",
     "pretty": "prettier --write src/",
     "start": "next start"
@@ -36,7 +38,6 @@
     "@eslint/eslintrc": "3.3.5",
     "@eslint/js": "9.39.4",
     "@philihp/prettier-config": "1.0.0",
-    "@types/jsonwebtoken": "9.0.10",
     "@types/node": "22.13.10",
     "@types/ramda": "0.31.1",
     "@types/react": "19.2.14",
@@ -47,7 +48,6 @@
     "lint-staged": "16.1.2",
     "prettier": "3.6.2",
     "supabase": "2.31.8",
-    "tsx": "4.22.0",
     "typescript": "5.8.3",
     "typescript-eslint": "8.59.3"
   }
