+++ src/app/character/sso.ts
@@ -1,8 +1,8 @@
-import SingleSignOn from '../../sso'
+import SingleSignOn from '../../sso.js'
 
-const EVE_CLIENT_ID = process.env.EVE_CLIENT_ID!
-const EVE_SECRET_KEY = process.env.EVE_SECRET_KEY!
-const EVE_CALLBACK_URL = process.env.EVE_CALLBACK_URL!
+const EVE_CLIENT_ID = process.env.EVE_CLIENT_ID
+const EVE_SECRET_KEY = process.env.EVE_SECRET_KEY
+const EVE_CALLBACK_URL = process.env.EVE_CALLBACK_URL
 
 export const scopes = [
   'publicData',
