--- src/hourly.ts
@@ -1,9 +1,0 @@
-import { authenticate, sudoSupabase } from './supabase'
-
-const execute = async () => {
-  await authenticate()
-  const res = await sudoSupabase.from('character_token').select('id, name')
-  console.log(res)
-}
-
-execute()
